Some warning cleanup

+++ src/mpid/ch4/netmod/ofi/am_impl.h
@@ -280,7 +280,7 @@ static inline int MPIDI_netmod_ofi_do_send_am_hdr(int                         ra
         msg_hdr->context_id = use_comm->context_id;
         msg_hdr->src_rank   = use_comm->rank;
     }
-    MPIU_Assert(use_comm->rank < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
+    MPIU_Assert((uint64_t)use_comm->rank < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
 
     AMREQ_OFI_HDR(sreq, clientdata).pack_buffer = NULL;
     MPID_cc_incr(sreq->cc_ptr, &c);
@@ -328,7 +328,7 @@ static inline int MPIDI_netmod_ofi_send_am_long(int           rank,
     MPIU_Assert(am_hdr_sz        < (1ULL << MPIDI_CH4_NMI_OFI_AM_HDR_SZ_BITS));
     MPIU_Assert(data_sz          < (1ULL << MPIDI_CH4_NMI_OFI_AM_DATA_SZ_BITS));
     MPIU_Assert(comm->context_id < (1    << MPIDI_CH4_NMI_OFI_AM_CONTEXT_ID_BITS));
-    MPIU_Assert(comm->rank       < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
+    MPIU_Assert((uint64_t)comm->rank       < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
 
     msg_hdr             = &AMREQ_OFI_HDR(sreq, msg_hdr);
     msg_hdr->handler_id = handler_id;
@@ -345,7 +345,7 @@ static inline int MPIDI_netmod_ofi_send_am_long(int           rank,
     /* Always allocates RMA ID from COMM_WORLD as the actual associated communicator
        is not available here */
     index = MPIDI_OFI_Index_allocator_alloc(COMM_OFI(MPIR_Process.comm_world).rma_id_allocator);
-    MPIU_Assert(index < MPIDI_Global.max_huge_rmas);
+    MPIU_Assert((int)index < MPIDI_Global.max_huge_rmas);
     lmt_info->rma_key = index << MPIDI_Global.huge_rma_shift;
 
     MPID_cc_incr(sreq->cc_ptr, &c); /* send completion */
@@ -400,11 +400,11 @@ static inline int MPIDI_netmod_ofi_send_am_short(int           rank,
     MPIDI_STATE_DECL(MPID_STATE_NETMOD_OFI_SEND_AM_SHORT);
     MPIDI_FUNC_ENTER(MPID_STATE_NETMOD_OFI_SEND_AM_SHORT);
 
-    MPIU_Assert(handler_id       < (1    << MPIDI_CH4_NMI_OFI_AM_HANDLER_ID_BITS));
-    MPIU_Assert(am_hdr_sz        < (1ULL << MPIDI_CH4_NMI_OFI_AM_HDR_SZ_BITS));
-    MPIU_Assert(count            < (1ULL << MPIDI_CH4_NMI_OFI_AM_DATA_SZ_BITS));
-    MPIU_Assert(comm->context_id < (1    << MPIDI_CH4_NMI_OFI_AM_CONTEXT_ID_BITS));
-    MPIU_Assert(comm->rank       < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
+    MPIU_Assert(handler_id           < (1    << MPIDI_CH4_NMI_OFI_AM_HANDLER_ID_BITS));
+    MPIU_Assert(am_hdr_sz            < (1ULL << MPIDI_CH4_NMI_OFI_AM_HDR_SZ_BITS));
+    MPIU_Assert((uint64_t)count      < (1ULL << MPIDI_CH4_NMI_OFI_AM_DATA_SZ_BITS));
+    MPIU_Assert(comm->context_id     < (1    << MPIDI_CH4_NMI_OFI_AM_CONTEXT_ID_BITS));
+    MPIU_Assert((uint64_t)comm->rank < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
 
     msg_hdr = &AMREQ_OFI_HDR(sreq, msg_hdr);
     msg_hdr->handler_id = handler_id;
@@ -553,7 +553,7 @@ static inline int MPIDI_netmod_do_inject(int           rank,
         msg_hdr.context_id = use_comm->context_id;
         msg_hdr.src_rank   = use_comm->rank;
     }
-    MPIU_Assert(use_comm->rank < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
+    MPIU_Assert((uint64_t)use_comm->rank < (1ULL << MPIDI_CH4_NMI_OFI_AM_RANK_BITS));
 
     msg_iov[0].iov_base = (void *) &msg_hdr;
     msg_iov[0].iov_len  = sizeof(msg_hdr);
